lang: backend: Builtin args are no longer reversed
Previously, args were passed to builtin funcitions in reverse order, ie
```pusl
print("foo", "bar")
```
would be passed to the print builtin with the String "bar" first, then
the string "foo". This doesn't make any sense and it is in the much more
reasonable order ["foo", "bar"] in this commit.

Signed-off-by: Samuel O'Brien <sam.obrien00@gmail.com>