lang: backend: Restructure pusl Function Type
Previously, the Function type was generic over an extra resolved field,
which was empty for serialization and was used to store extra runtime
data. This seemed like a good idea at the time, but as the project
progressed, it became unwieldy. This commit changes the structure so now
a ResolvedFunction (aka a function whose imports have been resolved) now
contains a Function, rather than mirroring the fields of one.

Signed-off-by: Samuel O'Brien <sam.obrien00@gmail.com>